+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## 2021-05-29
+### Added
+* Program can now discover if a DICOM file set has RT Struct, CT Image, RT Dose, or RT Plan elements. The presence of each is printed to the command line. The program can also determine which of these each individual DICOM file contains.
+
 ## 2021-05-28
 ### Added
 * Program can now find all DICOM files inside the program's default directory.